c++ moduleS

+++ cpp/code/testing.cpp
@@ -1,5 +1,5 @@
-#include <iostream>
 import Vector;
+import <iostream>;
 using namespace std;
 
 int main(int argc, char **argv) {
@@ -8,7 +8,7 @@ int main(int argc, char **argv) {
     const Vector vec(n);
     for (int i = 0; i < n; i++)
         cin >> vec[i];
-    int sum = 0;
+    double sum = 0;
     for (int i = 0; i < n; i++)
         sum += vec[i];
     cout << sum << endl;
